корректировки. Изменил название пкаетов в проекте. groupId и artifactId
