--- src/main/java/praktikum/IngredientType.java
@@ -1,11 +1,0 @@
-package praktikum;
-
-/**
- * Перечисление с типами ингредиентов.
- * SAUCE – соус
- * FILLING – начинка
- */
-public enum IngredientType {
-    SAUCE,
-    FILLING
-}
