+++ src/main/java/site/nomoreparties/stellarburgers/Bun.java
@@ -1,0 +1,25 @@
+package site.nomoreparties.stellarburgers;
+
+/**
+ * Модель булочки для бургера.
+ * Булочке можно дать название и назначить цену.
+ */
+public class Bun {
+
+    public String name;
+    public float price;
+
+    public Bun(String name, float price) {
+        this.name = name;
+        this.price = price;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public float getPrice() {
+        return price;
+    }
+
+}
