+++ src/main/java/site/nomoreparties/stellarburgers/IngredientType.java
@@ -1,0 +1,11 @@
+package site.nomoreparties.stellarburgers;
+
+/**
+ * Перечисление с типами ингредиентов.
+ * SAUCE – соус
+ * FILLING – начинка
+ */
+public enum IngredientType {
+    SAUCE,
+    FILLING
+}
