+++ src/test/java/site/nomoreparties/stellarburgers/BunTest.java
@@ -1,4 +1,4 @@
-package praktikum;
+package site.nomoreparties.stellarburgers;
 
 
 import org.junit.Test;
