+++ src/test/java/site/nomoreparties/stellarburgers/BurgerTest.java
@@ -1,4 +1,4 @@
-package praktikum;
+package site.nomoreparties.stellarburgers;
 
 
 import org.junit.Assert;
