+++ src/test/java/site/nomoreparties/stellarburgers/IngredientTest.java
@@ -1,4 +1,4 @@
-package praktikum;
+package site.nomoreparties.stellarburgers;
 
 import org.junit.Test;
 import static org.junit.Assert.assertEquals;
