+++ src/test/java/site/nomoreparties/stellarburgers/IngredientTypeTest.java
@@ -1,4 +1,4 @@
-package praktikum;
+package site.nomoreparties.stellarburgers;
 
 import org.junit.Assert;
 import org.junit.Test;
